+++ src/app/components/index.tsx
@@ -1,0 +1,4 @@
+import { SideNav } from "./SideNav";
+
+//
+export { SideNav }
