+++ src/app/content/connectors/page.tsx
@@ -1,0 +1,9 @@
+'use client'
+import React from 'react';
+//
+
+export default function APIIndex() {
+  return (
+    <>Connectors page</>
+  )
+}
