+++ src/app/content/crawlers/page.tsx
@@ -1,0 +1,10 @@
+'use client'
+import React from 'react';
+//
+
+export default function APIIndex() {
+  return (
+    <>Crawlers page</>
+  )
+}
+
