+++ src/app/page.tsx
@@ -1,19 +1,17 @@
-'use client';
+'use client'
 import React from 'react';
-import { useRouter } from 'next/router';
-
-import { RootLayout } from '@/components';
+//
+import { useRouter } from 'next/navigation';
+//
 import {
-  EuiButton,
-  EuiText,
-  EuiTitle,
   EuiFlexGroup,
   EuiFlexItem,
+  EuiTitle,
   EuiPanel,
+  EuiText,
   EuiSpacer,
-  EuiStat,
-  EuiHorizontalRule,
-} from "@elastic/eui";
+  EuiButton,
+} from '@elastic/eui';
 
 const GetStartedPanel = ({ heading, description, slug }) => {
   const router = useRouter();
@@ -30,16 +28,9 @@ const GetStartedPanel = ({ heading, description, slug }) => {
   )
 }
 
-export default function GettingStarted() {
-
-
-  const DATA = {
-    pageTitle: "Get started with Elasticsearch",
-    rightSideItems: [
-    ]
-  }
+export default function Home() {
   return (
-    <RootLayout pageData={DATA}>
+    <>
       <EuiFlexGroup direction="column">
         <EuiFlexItem><EuiTitle size="s"><h3>Ingest your content</h3></EuiTitle></EuiFlexItem>
         <EuiFlexGroup>
@@ -63,8 +54,7 @@ export default function GettingStarted() {
           </EuiFlexItem>
         </EuiFlexGroup>
       </EuiFlexGroup>
+    </>
 
-
-    </RootLayout>
   )
 }
