+++ src/utils/custom-eui-provider.tsx
@@ -1,10 +1,11 @@
 'use client';
-
-import { EuiProvider } from "@elastic/eui";
+import React from 'react';
+import { EuiProvider, EuiProviderProps } from "@elastic/eui";
+//
 const CustomEuiProvider = (props: Partial<EuiProviderProps<{}>>) => (
   <EuiProvider
     {...props}
   />
 );
 
-export default CustomEuiProvider;
+export { CustomEuiProvider };
