+++ src/utils/emotion-cache-provider.tsx
@@ -1,5 +1,6 @@
 'use client';
-
+import React from 'react';
+//
 import createCache, { EmotionCache, Options } from '@emotion/cache';
 import { useServerInsertedHTML } from 'next/navigation';
 import { ComponentType, PropsWithChildren, useMemo } from 'react';
